cabgo/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
python-dotenv==1.0.0
Flask-Migrate==4.0.5 # Optional, for database migrations (Step 4)
PyJWT==2.8.0         # For JWT authentication (Step 5)
# Add other dependencies as needed
